{"schema":"decodex/commit/1","summary":"Split progressive search details modules","authority":"manual"}
